getting started on the module structure.
Trying to use the yet to be written Oncoprint.

1. I'm going to need to setup a build too (gulp) sooner than I thought.
For now there is this temporary index.html file. Fine. But we need to
start using `require` like...yesterday.

2. Not exactly back to the drawing board but need to think more about
how to expose rendering callbacks.

3. I'm happy with LightTable setup, running `node http_server` to serve
up a page. Maybe this warrants some sort of "dev build" in the future.

Renderers will be a map of datatype -> renderer. This leaves the
computing of datatype up to the user. The other option would be to allow
users to specify a function that would compute the datatype. This
doesn't seem like the right move though. Too complex.

beginnings of a DOM structure.

puttings the rendering code in one main function for now. When it works
and seems to reasonably validate the design the core module I will break
it into another file and leave the core as the simple shell that is
should be. At this time it will be necessary to work on the build stuff.

Proof of concept? Done.