don't close over rows.
Rows are bound directly to the container element. Right now this is done
with "raw d3," so it could be wrapped in some helper functions, but the
point remains: do not keep two copies of the data, one in a clojure
within an oncoprint object, and one in the DOM. Keep one version of the
data, in one place: the DOM.

This cleans up a lot of the code too!