good old "memo sort"
This is a special case of a radix sort and is implemented as such. I use
radix sort to sort a list of indicies and reindex (reorder) each row
according to these indicies. This is not highly optimized, but it is
fairly clear and decoupled code.

Additionally, I introduce index functions for the radix sort. An index
function takes an element in the list to be sorted and returns the value
that should be passed to the metric function. They correspond to the
order of the genes that the radix sort uses to perform its sorting.