Some naming changes and cleaning up
Added 'D3SVGRuleset#apply' for ease of use/modularity

Factored out percentage conversion into 'D3SVGRule#_toLocalCoords'.
This may still not be the right place for it...

Removed 'D3SVGRule#filterData' and put its functionality into 'D3SVGRule#apply'

Changed 'id_member' to 'datum_id' which is now a function that takes
a datum and returns its id. This takes the place of what was
'D3SVGCellRenderer#data_key'

Moved the binding of the event cellRenderer.updateCells() into the cellRenderer
instead of Track where it was.

Changed 'renderX' to 'init' in track and cell renderer