Refactor; removing Track class and changing way rulesets are used
Changing the code to make RuleSets unstateful, so you render
using a ruleset. All the rendering now takes place in the oncoprint
renderer, no unnecessary delegation to a track renderer. Furthermore,
realizing that the track doesn't actually do any computation,
tracks are now simple objects that belong to the oncoprint and are
handled by the oncoprint, there is no need to break them into a different class.

This new paradigm is in progress, with gender track correctly rendering. Old
code has not yet been excised but it will be soon.