Add a separate track param for preformatted labels
Commit 6310fa4e508101097affb71ac7fc8b037da170bb introduced an option to
specify an HTML string for custom rendering of the label in the tooltip,
but it required passing the label as a non-primitive String object.
Since these may not behave identical to plain strings in all contexts,
this may lead to bugs in the long run, and it's safer to remove this
part of the API right away and simply add another track param instead.